+++ aprofiles/detection/foc.py
@@ -2,13 +2,13 @@
 # @desc A-Profiles - Fog or Condensation detection
 
 from aprofiles import utils
+import numpy as np
+from typing import Literal
 
-
-def _detect_fog_from_cloud_base_height(profiles, zmin_cloud):
+def _detect_fog_from_cloud(profiles, zmin_cloud):
     # returns a bool list with True where fog/condensation cases
     # if the base of the first cloud (given by the constructor) is below
-    first_cloud_base_height = profiles.data.cloud_base_height.data[:, 0]
-    # condition
+    first_cloud_base_height = profiles._get_lowest_clouds()
     foc = [True if x <= zmin_cloud else False for x in first_cloud_base_height]
     return foc
 
@@ -26,7 +26,7 @@ def _detect_fog_from_snr(profiles, z_snr, var, min_snr):
     foc = [True if x <= min_snr else False for x in snr]
     return foc
 
-def detect_foc(profiles, method="cloud_base", var="attenuated_backscatter_0", z_snr=2000., min_snr=2., zmin_cloud=200.):
+def detect_foc(profiles, method: Literal["cloud_base", "snr"]="cloud_base", var="attenuated_backscatter_0", z_snr=2000., min_snr=2., zmin_cloud=200.):
     """Detects fog or condensation.
 
     Args:
@@ -59,9 +59,9 @@ def detect_foc(profiles, method="cloud_base", var="attenuated_backscatter_0", z_
         ![Fog or condensation (foc) detection](../../assets/images/foc.png)
     """
 
-    if method == "cloud_base":
-        foc = _detect_fog_from_cloud_base_height(profiles, zmin_cloud)
-    elif method.upper() == "SNR":
+    if method == "cloud_base" and "clouds" in profiles.data:
+        foc = _detect_fog_from_cloud(profiles, zmin_cloud)
+    else:
         foc = _detect_fog_from_snr(profiles, z_snr, var, min_snr)
 
     # creates dataarray